refactor(role): change role import and model
