+++ app/Controllers/Http/AuthController.ts
@@ -1,5 +1,5 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
-import Roles from 'App/Enums/Roles'
+import Roles from 'App/constants/roles'
 import Role from 'App/Models/Role'
 import User from 'App/Models/User'
 import RegisterValidator from 'App/Validators/Auth/RegisterValidator'
@@ -14,16 +14,25 @@ export default class AuthController {
     return token.toJSON()
   }
 
+  /**
+   * User will be create with a default USER role
+   */
   public async register({ response, request, auth }: HttpContextContract) {
-    const data = await request.validate(RegisterValidator)
-    const user = await User.create({
-      ...data,
-      roleId: (await Role.findByOrFail('name', Roles.USER)).id,
-    })
-    const token = await auth.use('api').attempt(data.email, data.password, {
-      expiresIn: '10 days',
-    })
-    response.json({ success: true, user, token })
+    try {
+      const data = await request.validate(RegisterValidator)
+      const user = await User.create({
+        ...data,
+        roleId: (await Role.findByOrFail('name', Roles.USER)).id,
+      })
+      const token = await auth.use('api').attempt(data.email, data.password, {
+        expiresIn: '10 days',
+      })
+      response.json({ success: true, user, token })
+    } catch (error) {
+      // TODO add message or some redirect
+      console.log(error)
+      response.json({ success: false })
+    }
   }
 
   public async logout({ response, auth }: HttpContextContract) {
