+++ database/factories/UserFactory.ts
@@ -5,6 +5,7 @@ export default Factory.define(User, ({ faker }) => {
   return {
     email: faker.internet.email(),
     password: faker.internet.password(),
-    name: faker.random.word(),
+    name: faker.internet.userName(),
+    // TODO set default role
   }
 }).build()
